Code a Discord Bot/requirements.txt
discord.py
streamlit
python-dotenv
